+++ HISTORY.rst
@@ -9,6 +9,7 @@ Unreleased / master
 ------------------
 * New ``ispyb.sqlalchemy`` module containing `SQLAlchemy <https://www.sqlalchemy.org>`_ ORM models to interface with ISPyB
 * New ``ispyb.last_data_collections_on`` command-line interface to list the most recent data collections on the given beamlines
+* With the introduction of the SQLAlchemy models the existing ``ispyb.model`` has been deprecated and will not be developed further.
 
 5.9.1 (2021-01-28)
 ------------------